+++ src/observe.js
@@ -4,7 +4,7 @@
  * A lightweight, and easy-to-use, JavaScript library for observing any kind of JS object, or HTML DOM element, to detect changes.
  * 
  * @file        observe.js
- * @version     v0.3.0
+ * @version     v0.4.0
  * @author      Bunoon
  * @license     MIT License
  * @copyright   Bunoon 2023
@@ -28,7 +28,7 @@
         _configuration = {},
 
         // Variables: Attribute Names
-        _attribute_Name_Options = "data-observe-options";
+        _attribute_Name_Watch_Options = "data-observe-watch-options";
 
 
     /*
@@ -57,8 +57,8 @@
     function collectDOMObject( element ) {
         var result = true;
 
-        if ( isDefined( element ) && element.hasAttribute( _attribute_Name_Options ) ) {
-            var bindingOptionsData = element.getAttribute( _attribute_Name_Options );
+        if ( isDefined( element ) && element.hasAttribute( _attribute_Name_Watch_Options ) ) {
+            var bindingOptionsData = element.getAttribute( _attribute_Name_Watch_Options );
 
             if ( isDefinedString( bindingOptionsData ) ) {
                 var bindingOptions = getObjectFromString( bindingOptionsData );
@@ -70,18 +70,20 @@
                         element.id = newGuid();
                     }
 
+                    element.removeAttribute( _attribute_Name_Watch_Options );
+
                     createWatch( element, bindingOptions, element.id );
 
                 } else {
                     if ( !_configuration.safeMode ) {
-                        console.error( "The attribute '" + _attribute_Name_Options + "' is not a valid object." );
+                        console.error( "The attribute '" + _attribute_Name_Watch_Options + "' is not a valid object." );
                         result = false;
                     }
                 }
 
             } else {
                 if ( !_configuration.safeMode ) {
-                    console.error( "The attribute '" + _attribute_Name_Options + "' has not been set correctly." );
+                    console.error( "The attribute '" + _attribute_Name_Watch_Options + "' has not been set correctly." );
                     result = false;
                 }
             }
@@ -107,13 +109,13 @@
                 watch = {};
 
             watch.options = watchOptions;
-            watch.domElementId = domElementId;
             watch.totalChanges = 0;
 
             if ( isDefinedString( domElementId ) ) {
                 var domElement = _parameter_Document.getElementById( domElementId );
 
                 if ( isDefined( domElement ) ) {
+                    watch.domElementId = domElementId;
                     watch.cachedObject = domElement.outerHTML;
                     watch.originalObject = domElement.outerHTML;
                 }
@@ -124,22 +126,25 @@
             }
 
             watch.timer = setInterval( function() {
-                var currentDateTime = new Date();
-
-                if ( !isDefinedDate( watchOptions.starts ) || currentDateTime >= watchOptions.starts ) {
-                    watchObjectForChanges( storageId );
-
-                    if ( isDefinedDate( watchOptions.expires ) && currentDateTime >= watchOptions.expires ) {
-                        cancelWatchObject( storageId );
-                    }
-                }
-
+                watchTimer( watchOptions, storageId );
             }, watchOptions.timeout );
 
             _watches[ storageId ] = watch;
         }
 
         return storageId;
+    }
+
+    function watchTimer( watchOptions, storageId ) {
+        var currentDateTime = new Date();
+
+        if ( !isDefinedDate( watchOptions.starts ) || currentDateTime >= watchOptions.starts ) {
+            watchObjectForChanges( storageId );
+
+            if ( isDefinedDate( watchOptions.expires ) && currentDateTime >= watchOptions.expires ) {
+                cancelWatchObject( storageId );
+            }
+        }
     }
 
     function watchObjectForChanges( storageId ) {
@@ -153,6 +158,8 @@
 
                 if ( isDefined( domElement ) ) {
                     watch.originalObject = domElement.outerHTML;
+                } else {
+                    watch.originalObject = _string.empty;
                 }
             }
 
@@ -161,9 +168,7 @@
                 originalObjectJson = !isDomElement ? JSON.stringify( originalObject ) : originalObject;
 
             if ( cachedObject !== originalObjectJson ) {
-                var watchOptions = watch.options;
-
-                if ( watchOptions.reset ) {
+                if ( watch.options.reset ) {
                     if ( isDomElement ) {
                         domElement.outerHTML = watch.cachedObject;
                     } else {
@@ -175,30 +180,30 @@
                 }
 
                 if ( isDomElement ) {
-                    fireCustomTrigger( watchOptions.onChange, cachedObject, originalObjectJson );
+                    fireCustomTrigger( watch.options.onChange, cachedObject, originalObjectJson );
                 } else {
 
                     var oldValue = getObjectFromString( cachedObject ).result,
                         newValue = getObjectFromString( originalObjectJson ).result;
 
-                    fireCustomTrigger( watchOptions.onChange, oldValue, newValue );
+                    fireCustomTrigger( watch.options.onChange, oldValue, newValue );
 
-                    if ( isDefinedFunction( watchOptions.onPropertyChange ) && !isDefinedArray( oldValue ) ) {
-                        compareWatchObjectProperties( oldValue, newValue, watchOptions );
+                    if ( isDefinedFunction( watch.options.onPropertyChange ) && !isDefinedArray( oldValue ) ) {
+                        compareWatchObjectProperties( oldValue, newValue, watch.options );
                     }
                 }
 
-                if ( watchOptions.pauseTimeoutOnChange > 0 ) {
-                    pauseWatchObject( storageId, watchOptions.pauseTimeoutOnChange );
+                if ( watch.options.pauseTimeoutOnChange > 0 ) {
+                    pauseWatchObject( storageId, watch.options.pauseTimeoutOnChange );
                 }
 
-                if ( watchOptions.cancelOnChange ) {
+                if ( watch.options.cancelOnChange ) {
                     cancelWatchObject( storageId );
                 }
 
                 watch.totalChanges++;
 
-                if ( watchOptions.maximumChangesBeforeCanceling > 0 && watch.totalChanges >= watchOptions.maximumChangesBeforeCanceling ) {
+                if ( watch.options.maximumChangesBeforeCanceling > 0 && watch.totalChanges >= watch.options.maximumChangesBeforeCanceling ) {
                     cancelWatchObject( storageId );
                 }
             }
@@ -237,11 +242,9 @@
 
     function cancelWatchObject( storageId ) {
         if ( _watches.hasOwnProperty( storageId ) ) {
-            var watchOptions = _watches[ storageId ].options;
-
-            fireCustomTrigger( watchOptions.onCancel, storageId );
-
+            fireCustomTrigger( _watches[ storageId ].options.onCancel, storageId );
             clearTimeout( _watches[ storageId ].timer );
+            
             delete _watches[ storageId ];
         }
     }
@@ -457,7 +460,7 @@
      */
 
     /**
-     * watchObject().
+     * watch().
      * 
      * Adds an object that should be watched for changes.
      * 
@@ -468,7 +471,7 @@
      * 
      * @returns     {string}                                                The ID that object watch is stored under.
      */
-    this.watchObject = function( object, options ) {
+    this.watch = function( object, options ) {
         return createWatch( object, options );
     };
 
@@ -592,6 +595,21 @@
         return result;
     };
 
+    /**
+     * searchDomForNewWatches().
+     * 
+     * Searches the DOM for new elements to watch, and adds them.
+     * 
+     * @public
+     * 
+     * @returns     {Object}                                                The Observe.js class instance.
+     */
+    this.searchDomForNewWatches = function() {
+        collectDOMObjects();
+
+        return this;
+    };
+
 
     /*
      * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
@@ -640,7 +658,7 @@
      * @returns     {string}                                                The version number.
      */
     this.getVersion = function() {
-        return "0.3.0";
+        return "0.4.0";
     };